Dialer: Proximity Speakerphone originally done by dankoman30 rewritten and moved to dialer by beanstown106
PureNexus edits:
*moved settings from TelephonyService to Dialer
*Made a "proximity speakerphone" fragment in dialer settings
*Changed proximity delay from slimseekbarpreference to a ListPreference
*Updated strings

The main reason behind the move from TelephonyService to Dialer was because in the call settings menu google would add more menu's after our preference category which looked really off. changed from seekbar to a listpreference for a more consistent look as well as a better UI for the end user.

Authorship is dankomans as this was his mod originally i just rewrote the setting aspect of it

credits to @ezio84 for forward porting to 7.0

Change-Id: I6ea37ef37ec48f965ec550d189ff26239381083d